Move around the serverspec stuff, standardise
