+++ test/integration/general/serverspec/install_netdata_spec.rb
@@ -1,11 +1,11 @@
-require "#{ENV['BUSSER_ROOT']}/../kitchen/data/serverspec_helper"
+require 'spec_helper'
 
 describe 'Netdata installation' do
-  it 'has Netdata service port listening' do
-    expect(port(19_999)).to be_listening
+  describe port(19_999) do
+    it { should be_listening }
   end
 
-  it 'has Netdata process running' do
-    expect(process('netdata')).to be_running
+  describe process('netdata') do
+    it { should be_running }
   end
 end
